refactor(CharData): change directory and initiate from controller
