+++ frontend/lib/app/data/models/chart_data.dart
